[refactor]move qlens set and get function to metadatabuiler
