--- vllm_ascend/_310p/attention/metadata.py
@@ -1,36 +1,0 @@
-#
-# Copyright (c) 2026 Huawei Technologies Co., Ltd. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# This file is a part of the vllm-ascend project.
-#
-
-"""Dynamic 310P fields attached to AscendMetadata at runtime."""
-
-import torch
-
-from vllm_ascend.attention.attention_v1 import AscendMetadata
-
-QUERY_LENS_CPU_ATTR = "query_lens_cpu"
-
-
-def set_query_lens_cpu(attn_metadata: AscendMetadata, query_lens_cpu: torch.Tensor) -> None:
-    """Attach host qLens for ATB splitfuse without extending upstream AscendMetadata."""
-    setattr(attn_metadata, QUERY_LENS_CPU_ATTR, query_lens_cpu)
-
-
-def get_query_lens_cpu(attn_metadata: AscendMetadata) -> torch.Tensor | None:
-    value = getattr(attn_metadata, QUERY_LENS_CPU_ATTR, None)
-    if value is None:
-        return None
-    return value
