+++ vllm_ascend/_310p/attention/metadata_builder.py
@@ -25,13 +25,26 @@ from vllm_ascend._310p.attention.attention_mask import (
     AttentionMaskBuilder310,
     is_compressed_mask_supported,
 )
-from vllm_ascend._310p.attention.metadata import set_query_lens_cpu
 from vllm_ascend.attention.attention_v1 import (
     AscendAttentionMetadataBuilder,
     AscendAttentionState,
     AscendMetadata,
 )
 from vllm_ascend.attention.utils import AscendCommonAttentionMetadata
+
+QUERY_LENS_CPU_ATTR = "query_lens_cpu"
+
+
+def set_query_lens_cpu(attn_metadata: AscendMetadata, query_lens_cpu: torch.Tensor) -> None:
+    """Attach host qLens for ATB splitfuse without extending upstream AscendMetadata."""
+    setattr(attn_metadata, QUERY_LENS_CPU_ATTR, query_lens_cpu)
+
+
+def get_query_lens_cpu(attn_metadata: AscendMetadata) -> torch.Tensor | None:
+    value = getattr(attn_metadata, QUERY_LENS_CPU_ATTR, None)
+    if value is None:
+        return None
+    return value
 
 
 class AscendAttentionMetadataBuilder310(AscendAttentionMetadataBuilder):
